Rep-Counter-master/Rep-Counter-master/requirements.txt
tensorflow-gpu
scipy
pyyaml
opencv-python==3.4
Flask
flask-cors
